requirements.txt
# Core requirements
numpy<2.0.0  # Downgraded to solve compatibility issues
pandas
scikit-learn
dill

# ML Tracking
mlflow

# Visualization
matplotlib
seaborn
notebook

# LangChain Framework (if needed)
langchain
langchain-groq
langchain-community
langchain-chroma
langchain-text-splitters

# Vector Stores
chromadb
fastembed

# LLM Integrations
ollama
pyautogen
langchain-google-genai
langchain
langchain-groq
python-dotenv
# Utilities
python-dotenv
yfinance
pydantic
duckduckgo-search
newsapi-python